added database

+++ src/db_aux.py
@@ -1,9 +1,7 @@
 from models.db import engine
 from sqlalchemy.orm import relationship, backref, sessionmaker
 import json
-from models.Livros import Livros, Edicao
-from models.Autor import Autor
-from models.LivroAutor import LivroAutor
+from models.Livros import Livros, Edicao, LivroAutor, Autor
 from models.LivrosTemp import LivrosTemp
 
 
@@ -15,35 +13,54 @@ class DBAux():
         self.conn = engine.connect()
 
     def create_book(self, data):
-        author = self.find_or_create_author(data.autor)
-        book = Livros(titulo=data.titulo)
-        session.add(book)
-        session.commit()
+        try:
 
-        edition = self.create_edition(data.edicao, book)
-        livroautor = self.create_book_author(book, author)
-
-    def find_or_create_author(self, autor):
-        query = self.session.query(Autor).filter(Autor.nome == autor).first()
-
-        if len(query):
-            return query[0]
-        else:
-            autor = Autor(nome=autor)
-            self.session.add()
+            book = Livros(titulo=data['titulo'])
+            self.session.add(book)
             self.session.commit()
 
+            self.session.flush()
+        except Exception as e:
+            print(e)
+
+        author = self.create_author(data['autor'], book)
+
+        edition = self.create_edition(data['edicao'], data['ano'], book)
+
+    def create_author(self, author_nome, book):
+        try:
+            author = Autor(nome=author_nome)
+            self.session.add(author)
+            self.session.commit()
+            self.session.flush()
+            association = self.create_book_author(book, author)
+            return author
+        except Exception as e:
+            print(e)
+            print("Problema para criar o author")
+            return None
+
     def create_edition(self, edicao, ano, livro):
-        edicao = Edicao(numero=edicao, ano=ano, codigolivro=livro.codigo)
-        self.session.add(edicao)
-        self.session.commit()
-        return edicao
+        try:
+            edicao = Edicao(numero=edicao, ano=ano, codigolivro=livro.codigo)
+            self.session.add(edicao)
+            self.session.commit()
+            return edicao
+        except Exception as e:
+            print(e)
+            print("Problema ao criar edição")
+            return None
 
     def create_book_author(self, book, author):
-        livroautor = LivroAutor(codigolivro=book.codigo, codigoautor=autor.codigo)
-        self.session.add(livroautor)
-        self.session.commit()
-        return edicao
+        try:
+            livroauthor = LivroAutor(codigolivro=book.codigo, codigoautor=author.codigo)
+            self.session.add(livroauthor)
+            self.session.commit()
+            return livroauthor
+        except Exception as e:
+            print(e)
+            print("Problema ao criar associação book_author")
+            return None
 
     def find_book(self, book_title, book_author):
         list_book = []
@@ -82,9 +99,12 @@ class DBAux():
         return json.dumps(list_book)
 
     def delete_book(self, title):
-        obj = self.session.query(Livros).filter(Livros.titulo.like('%' + title.titulo)).first()
-        self.session.delete(obj)
-        self.session.commit()
+        obj = self.session.query(Livros)\
+            .filter(Livros.titulo.ilike('%' + title['titulo'] + '%')).first()
+
+        if obj:
+            self.session.delete(obj)
+            self.session.flush()
 
     def update_book(self, data):
         print(data)
@@ -98,7 +118,7 @@ class DBAux():
         edition.ano = data['ano']
         self.session.commit()
 
-        autor = self.session.query(Autor)\
+        author = self.session.query(Autor)\
             .filter(LivroAutor.codigoautor == Autor.codigo, LivroAutor.codigolivro == data['codigo']).first()
-        autor.nome = data['autor']
+        author.nome = data['autor']
         self.session.commit()
